Editor: Move and unify the inserter and list view states
