+++ packages/edit-post/src/store/selectors.js
@@ -13,13 +13,13 @@ import { store as coreStore } from '@wordpress/core-data';
 import { store as editorStore } from '@wordpress/editor';
 import deprecated from '@wordpress/deprecated';
 
+/**
+ * Internal dependencies
+ */
+import { unlock } from '../lock-unlock';
+
 const EMPTY_ARRAY = [];
 const EMPTY_OBJECT = {};
-const EMPTY_INSERTION_POINT = {
-	rootClientId: undefined,
-	insertionIndex: undefined,
-	filterValue: undefined,
-};
 
 /**
  * Returns the current editing mode.
@@ -487,28 +487,41 @@ export const __experimentalGetPreviewDeviceType = createRegistrySelector(
 /**
  * Returns true if the inserter is opened.
  *
+ * @deprecated
+ *
  * @param {Object} state Global application state.
  *
  * @return {boolean} Whether the inserter is opened.
  */
-export function isInserterOpened( state ) {
-	return !! state.blockInserterPanel;
-}
+export const isInserterOpened = createRegistrySelector( ( select ) => () => {
+	deprecated( `select( 'core/edit-post' ).isInserterOpened`, {
+		since: '6.5',
+		alternative: `select( 'core/editor' ).isInserterOpened`,
+	} );
+	return select( editorStore ).isInserterOpened();
+} );
 
 /**
  * Get the insertion point for the inserter.
+ *
+ * @deprecated
  *
  * @param {Object} state Global application state.
  *
  * @return {Object} The root client ID, index to insert at and starting filter value.
  */
-export function __experimentalGetInsertionPoint( state ) {
-	if ( typeof state.blockInserterPanel === 'boolean' ) {
-		return EMPTY_INSERTION_POINT;
+export const __experimentalGetInsertionPoint = createRegistrySelector(
+	( select ) => () => {
+		deprecated(
+			`select( 'core/edit-post' ).__experimentalGetInsertionPoint`,
+			{
+				since: '6.5',
+				version: '6.7',
+			}
+		);
+		return unlock( select( editorStore ) ).getInsertionPoint();
 	}
-
-	return state.blockInserterPanel;
-}
+);
 
 /**
  * Returns true if the list view is opened.
@@ -517,9 +530,13 @@ export function __experimentalGetInsertionPoint( state ) {
  *
  * @return {boolean} Whether the list view is opened.
  */
-export function isListViewOpened( state ) {
-	return state.listViewPanel;
-}
+export const isListViewOpened = createRegistrySelector( ( select ) => () => {
+	deprecated( `select( 'core/edit-post' ).isListViewOpened`, {
+		since: '6.5',
+		alternative: `select( 'core/editor' ).isListViewOpened`,
+	} );
+	return select( editorStore ).isListViewOpened();
+} );
 
 /**
  * Returns true if the template editing mode is enabled.
