+++ packages/edit-site/src/components/block-editor/use-site-editor-settings.js
@@ -2,7 +2,7 @@
  * WordPress dependencies
  */
 import { useViewportMatch } from '@wordpress/compose';
-import { useDispatch, useSelect } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
 import { store as coreStore } from '@wordpress/core-data';
 import { privateApis as editorPrivateApis } from '@wordpress/editor';
@@ -89,7 +89,6 @@ function useArchiveLabel( templateSlug ) {
 }
 
 export function useSpecificEditorSettings() {
-	const { setIsInserterOpened } = useDispatch( editSiteStore );
 	const isLargeViewport = useViewportMatch( 'medium' );
 	const {
 		templateSlug,
@@ -152,7 +151,6 @@ export function useSpecificEditorSettings() {
 			...settings,
 
 			supportsTemplateMode: true,
-			__experimentalSetIsInserterOpened: setIsInserterOpened,
 			focusMode: canvasMode === 'view' && focusMode ? false : focusMode,
 			allowRightClickOverrides,
 			isDistractionFree,
@@ -166,7 +164,6 @@ export function useSpecificEditorSettings() {
 		};
 	}, [
 		settings,
-		setIsInserterOpened,
 		focusMode,
 		allowRightClickOverrides,
 		isDistractionFree,
