+++ packages/edit-site/src/components/keyboard-shortcuts/edit-mode.js
@@ -6,6 +6,7 @@ import { useDispatch, useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { store as blockEditorStore } from '@wordpress/block-editor';
 import { store as interfaceStore } from '@wordpress/interface';
+import { store as editorStore } from '@wordpress/editor';
 import { createBlock } from '@wordpress/blocks';
 
 /**
@@ -18,7 +19,7 @@ import { STORE_NAME } from '../../store/constants';
 function KeyboardShortcutsEditMode() {
 	const { getEditorMode } = useSelect( editSiteStore );
 	const isListViewOpen = useSelect(
-		( select ) => select( editSiteStore ).isListViewOpened(),
+		( select ) => select( editorStore ).isListViewOpened(),
 		[]
 	);
 	const isBlockInspectorOpen = useSelect(
@@ -29,11 +30,11 @@ function KeyboardShortcutsEditMode() {
 		[]
 	);
 	const { redo, undo } = useDispatch( coreStore );
-	const { setIsListViewOpened, switchEditorMode, toggleDistractionFree } =
+	const { switchEditorMode, toggleDistractionFree } =
 		useDispatch( editSiteStore );
 	const { enableComplementaryArea, disableComplementaryArea } =
 		useDispatch( interfaceStore );
-
+	const { setIsListViewOpened } = useDispatch( editorStore );
 	const { replaceBlocks } = useDispatch( blockEditorStore );
 	const { getBlockName, getSelectedBlockClientId, getBlockAttributes } =
 		useSelect( blockEditorStore );
