+++ packages/edit-site/src/store/test/actions.js
@@ -76,34 +76,6 @@ describe( 'actions', () => {
 		} );
 	} );
 
-	describe( 'setIsListViewOpened', () => {
-		it( 'should set the list view opened state', () => {
-			const registry = createRegistryWithStores();
-
-			registry.dispatch( editSiteStore ).setIsListViewOpened( true );
-			expect( registry.select( editSiteStore ).isListViewOpened() ).toBe(
-				true
-			);
-
-			registry.dispatch( editSiteStore ).setIsListViewOpened( false );
-			expect( registry.select( editSiteStore ).isListViewOpened() ).toBe(
-				false
-			);
-		} );
-		it( 'should turn off distraction free mode when opening the list view', () => {
-			const registry = createRegistryWithStores();
-			registry
-				.dispatch( preferencesStore )
-				.set( 'core/edit-site', 'distractionFree', true );
-			registry.dispatch( editSiteStore ).setIsListViewOpened( true );
-			expect(
-				registry
-					.select( preferencesStore )
-					.get( 'core/edit-site', 'distractionFree' )
-			).toBe( false );
-		} );
-	} );
-
 	describe( 'openGeneralSidebar', () => {
 		it( 'should turn off distraction free mode when opening a general sidebar', () => {
 			const registry = createRegistryWithStores();
@@ -160,10 +132,10 @@ describe( 'actions', () => {
 					.select( preferencesStore )
 					.get( 'core/edit-site', 'fixedToolbar' )
 			).toBe( true );
-			expect( registry.select( editSiteStore ).isListViewOpened() ).toBe(
+			expect( registry.select( editorStore ).isListViewOpened() ).toBe(
 				false
 			);
-			expect( registry.select( editSiteStore ).isInserterOpened() ).toBe(
+			expect( registry.select( editorStore ).isInserterOpened() ).toBe(
 				false
 			);
 			expect(
