+++ packages/edit-site/src/store/test/selectors.js
@@ -10,8 +10,6 @@ import {
 	getCanUserCreateMedia,
 	getEditedPostType,
 	getEditedPostId,
-	isInserterOpened,
-	isListViewOpened,
 	isPage,
 } from '../selectors';
 
@@ -42,28 +40,6 @@ describe( 'selectors', () => {
 		it( 'returns the template type', () => {
 			const state = { editedPost: { postType: 'wp_template' } };
 			expect( getEditedPostType( state ) ).toBe( 'wp_template' );
-		} );
-	} );
-
-	describe( 'isInserterOpened', () => {
-		it( 'returns the block inserter panel isOpened state', () => {
-			const state = {
-				blockInserterPanel: true,
-			};
-			expect( isInserterOpened( state ) ).toBe( true );
-			state.blockInserterPanel = false;
-			expect( isInserterOpened( state ) ).toBe( false );
-		} );
-	} );
-
-	describe( 'isListViewOpened', () => {
-		it( 'returns the list view panel isOpened state', () => {
-			const state = {
-				listViewPanel: true,
-			};
-			expect( isListViewOpened( state ) ).toBe( true );
-			state.listViewPanel = false;
-			expect( isListViewOpened( state ) ).toBe( false );
 		} );
 	} );
